requirements.txt
Django==4.0.6
gunicorn==20.1.0
nltk==3.7
textblob==0.17.1
psycopg2==2.9.3
Pillow==9.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
word-forms==2.1.0
requests==2.31.0
textstat==0.7.3
gensim==4.3.2
openai==1.6.1
ffmpeg-python==0.2.0